output/quantum_engine_runner.py
#!/usr/bin/env python3
"""
GBM Quantum Pipeline — Google Quantum Engine Ready
===================================================
Carica su Google Cloud e lancia:
    gcloud auth login
    pip install cirq-google
    python quantum_engine_runner.py

Oppure su Google Colab:
    !pip install cirq-google
    # copia questo script nella prima cella
"""

import cirq
# import cirq_google  # decommentare per Google Quantum Engine

# ── Configurazione Google Quantum Engine ──
PROJECT_ID = "YOUR_GOOGLE_CLOUD_PROJECT_ID"  # ← inserisci qui
PROCESSOR  = "rainbow"  # oppure "weber", "willow"

def run_on_engine(circuit, repetitions=1000):
    """Esegui su hardware quantistico reale di Google."""
    # Decommentare per uso reale:
    # engine  = cirq_google.Engine(project_id=PROJECT_ID)
    # sampler = engine.sampler(processor_id=PROCESSOR)
    # return sampler.run(circuit, repetitions=repetitions)
    
    # Simulazione locale (fallback):
    simulator = cirq.Simulator()
    return simulator.run(circuit, repetitions=repetitions)

# ── Candidati GBM (dal ranking della pipeline) ──
CANDIDATES = [
    {"gene": "MDM2", "drug": "milademetan", "pocket": "p53_binding_cleft"},
    {"gene": "PIK3CA", "drug": "BKM120", "pocket": "ATP_site"},
    {"gene": "CDK4", "drug": "palbociclib", "pocket": "ATP_site"},
    {"gene": "EGFR", "drug": "erlotinib", "pocket": "ATP_site"},
]

# ── VQE per ogni candidato ──
from quantum_sim import QuantumPipelineRunner

runner  = QuantumPipelineRunner()
results = runner.run_all(CANDIDATES)

for r in results:
    print(f"\n{r.drug} → {r.gene}")
    print(f"  VQE Energy:    {r.vqe_energy:.4f} kcal/mol")
    print(f"  Binding Prob:  {r.binding_prob:.4f}")
    print(f"  Entanglement:  {r.entanglement:.4f}")
    print(f"  QAOA Score:    {r.qaoa_score:.4f}")
    print(f"  QML Score:     {r.qml_prediction:.4f}")
    print(f"  Circuit depth: {r.circuit_depth}")

# ── Deploy su Google Cloud ──
# gcloud run deploy gbm-quantum \
#   --source . \
#   --platform managed \
#   --region us-central1 \
#   --allow-unauthenticated
