output/requirements.txt
numpy>=1.24
pandas>=2.0
scipy>=1.10
matplotlib>=3.7
scikit-learn>=1.3
biopython>=1.81
cirq-core>=1.6
MDAnalysis>=2.6
requests>=2.31
